+++ library/core/engine/src/test/kotlin/de/lise/fluxflow/engine/step/StepActivationServiceTest.kt
@@ -121,6 +121,7 @@ class StepActivationServiceTest {
                 ),
                 DataDefinitionBuilder(
                     mock {},
+                    mock {},
                     mock {}
                 ),
                 mock {},