+++ library/core/engine/src/test/kotlin/de/lise/fluxflow/engine/step/data/DataTestStepAnnotation.kt
@@ -1,0 +1,6 @@
+package de.lise.fluxflow.engine.step.data
+
+import de.lise.fluxflow.stereotyped.metadata.Metadata
+
+@Metadata("key")
+annotation class DataTestStepAnnotation(val value: String)